+++ include/valuation.h
@@ -5,5 +5,5 @@ typedef enum {
     FALSE = 0, TRUE = 1, UNKNOWN = 2
 } valuation;
 
-valuation* initValuations(uint16_t nbVars);
+valuation* initValuations(uint32_t nbVars);
 void printValuation(valuation v);
