+++ src/valuation.c
@@ -1,8 +1,8 @@
 #include "valuation.h"
 
-valuation* initValuations(uint16_t nbVars){
+valuation* initValuations(uint32_t nbVars){
     valuation* ret = (valuation*)malloc(nbVars * sizeof(valuation));
-    for (uint16_t i = 0; i < nbVars; i++)
+    for (uint32_t i = 0; i < nbVars; i++)
         ret[i] = UNKNOWN;
     return ret;
 }
